ci: Changes to address lint errors
Most changes are minimal and reflect findings from the new golangci-lint
configuration.

+++ v2/pkg/server/server_test.go
@@ -27,7 +27,7 @@ func testGetDigit(ctx context.Context, t *testing.T, request *generated.GetDigit
 	if err != nil {
 		t.Errorf("Error calling FractionalDigit: %v", err)
 	}
-	if actual.Digit != uint32(expected) {
+	if uint64(actual.Digit) != expected {
 		t.Errorf("Checking index: %d: expected %d got %d", request.Index, expected, actual.Digit)
 	}
 }
